requirements.txt
langchain-community==0.0.10
langchain==0.0.350
langchain-openai==0.0.5
pinecone-client==5.0.1
python-dotenv==1.0.0
openai>=1.10.0
tiktoken==0.5.2
sentence-transformers==3.3.1
